# openapi_spec.yaml
...
info:
  description: Auto-generated OpenAPI spec for oqlos
  title: oqlos API
  version: 1.0.0
openapi: 3.0.3
paths:
...
  /api/status:
    get:
      operationId: status
      responses:
        '200': *id001
        '401': *id002
        '404': *id003
        '500': *id004
      summary: GET /api/status
      tags:
      - fastapi
      - status
...
  /api/v1/peripherals/{peripheral_id}:
    get:
      operationId: get_peripheral
      parameters:
      - in: path
        name: peripheral_id
        required: true
        schema:
          type: string
      - in: query
        name: peripheral_id
        required: false
        schema:
          type: str
      responses:
        '200': *id001
        '401': *id002
        '404': *id003
        '500': *id004
      summary: Get specific peripheral
      tags:
      - v1
      - fastapi
    put:
      operationId: update_peripheral
      parameters:
      - in: path
        name: peripheral_id
        required: true
        schema:
          type: string
      - in: query
        name: peripheral_id
        required: false
        schema:
          type: str
      requestBody:
        content:
          application/json:
            schema:
              properties:
                data:
                  type: object
                id:
                  type: string
                name:
                  type: string
              type: object
        required: true
      responses:
        '200': *id001
        '401': *id002
        '404': *id003
        '500': *id004
      summary: Update peripheral via PUT (for tests)
      tags:
      - v1
      - fastapi
...
  /api/v1/plugins/:
    get:
      operationId: list_plugins
      responses:
        '200': *id001
        '401': *id002
        '404': *id003
        '500': *id004
      summary: List all registered hardware plugins.
      tags:
      - v1
      - fastapi
  /api/v1/plugins/status:
    get:
      operationId: get_plugin_status
      responses:
        '200': *id001
        '401': *id002
        '404': *id003
        '500': *id004
      summary: Get overall status of all plugins.
      tags:
      - v1
      - fastapi
  /api/v1/plugins/validate:
    post:
      operationId: validate_plugin_configs
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        '201': *id005
        '400': *id006
        '401': *id002
        '404': *id003
        '500': *id004
      summary: Validate configurations for multiple plugins.
      tags:
      - v1
      - fastapi
  /api/v1/plugins/{plugin_id}:
    get:
      operationId: get_plugin_info
      parameters:
      - in: path
        name: plugin_id
        required: true
        schema:
          type: string
      - in: query
        name: plugin_id
        required: false
        schema:
          type: str
      responses:
        '200': *id001
        '401': *id002
        '404': *id003
        '500': *id004
      summary: Get information about a specific plugin.
      tags:
      - v1
      - fastapi
  /api/v1/plugins/{plugin_id}/connect:
    post:
      operationId: connect_plugin
      parameters:
      - in: path
        name: plugin_id
        required: true
        schema:
          type: string
      - in: query
        name: plugin_id
        required: false
        schema:
          type: str
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        '201': *id005
        '400': *id006
        '401': *id002
        '404': *id003
        '500': *id004
      summary: Connect to a hardware plugin.
      tags:
      - v1
      - fastapi
  /api/v1/plugins/{plugin_id}/disconnect:
    post:
      operationId: disconnect_plugin
      parameters:
      - in: path
        name: plugin_id
        required: true
        schema:
          type: string
      - in: query
        name: plugin_id
        required: false
        schema:
          type: str
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        '201': *id005
        '400': *id006
        '401': *id002
        '404': *id003
        '500': *id004
      summary: Disconnect from a hardware plugin.
      tags:
      - v1
      - fastapi
  /api/v1/plugins/{plugin_id}/execute:
    post:
      operationId: execute_plugin_command
      parameters:
      - in: path
        name: plugin_id
        required: true
        schema:
          type: string
      - in: query
        name: plugin_id
        required: false
        schema:
          type: str
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        '201': *id005
        '400': *id006
        '401': *id002
        '404': *id003
        '500': *id004
      summary: Execute a command on a hardware plugin.
      tags:
      - v1
      - fastapi
...
servers:
- description: Local development
  url: http://localhost:8101
- description: Relative
  url: /